TSG S2: avoid the possibility of the AI neutralizing villages...
...which it neither burns or captures.
The AI has two sides which are presented as one to the player. It's not the best
approach, because it is possible that the player realizes it looking at the side
informations, e.g. how many units the enemy side has listed there.

This commit unifies both sides, while still keeping the units from formerly side
3 being unable to burn villages, but, as now being in a side with a leader, they
can capture them.

An advantage of having two sides was, that the units could have different AI
settings, and that the leaderless side was not interrested in villages.
The AI settings will be adressed in a seperate commit.

(bug #3032)

[ci skip]