TSG S5: Duo leader setup for the zombies
This adds on the one hand a second leader with secons keep in the same caslte.
for the same side, on the other hand ich changes the speack when they die.
They adverturers are not assuming they killed the undead masters anymore if they
saw the lich before.

(mentioned in #1954)

[ci skip]

(cherry-picked from commit f1a465b3779c0194613971fbbcae4fe43a69dd26)