TSG: change inclusion of death events
That way less events need to be duplicated in scenarios.
It also does not include the full Hylas death event in S7a to S9a.
Previously the Jarek event was not present in one of the final scenarios,
now both include it, but the shorter variant of it.

[ci skip]

(cherry-picked from commit 29b4196556dbe401766f48d0f46faa6a9245d351)