Pofix fixes
The tool that generates .po files splits lines in order to limit line
length, and pofix can't track split strings. Thus, pofix search rules need
to be shorter so that pofix finds the affected strings.

I also moved the replacements to the right text domains, placed wesnoth-l
to the right spot, and removed the "Asheviere is the Queen of Wesnoth"
replacement because it isn't pofixable. (The second part of the string
becomes just "Queen of Wesnoth." and "queen of Wesnoth." with lowercase Q
is used legitimately elsewhere.)

[ci skip]