Micro AIs: simplify unit variable handling functions
The main reason for doing this, besides simplifying the code, is to avoid using the WML table dump __cfg, which is slow. This includes a change of the format in which the variables are stored, but since these are internal Micro AI utility functions, that does not matter.

(cherry-picked from commit 133568efe6851944f0b9ddaaf75ccaff0cad6a4d)