try to detect umc versions checks
now the game attepts to automatically detect 'good' umc code that check
for newer wesnoth versions before attempting to use new api
```
if wesnoth.compare_version(game_coinfig.version, ">=", "1.14.9") then
.. use new 1.14.9 api
end
```
and raise the api level in that case. of course this does not work
perfectly and wrongly detect version checks when when there are none, in
particular theere is no way to detect leaving that 'if' or whether there
is an 'if' in the first place, so player could in theory still use the
new api after they leave that 'if' . The better solution which is the
`wesnoth.switch_version` function  obviously does not work in for api
that is new in 1.14.6 becasue that function does not exists in 1.14.5 or
older. hopefully at some point, probably at wesnoth 1.15 we can remove
this hack so that people will use wesnoth.switch_version() instead.