show more scenarios in choose_level debug command
the old code didn't work well for mp campaigns that dynamicially set the
next scenario in endlevel. The new code just shows all scenarios that
belong to the same addon, since this is a debug command it better to sho
a little too much than not to show all scenarios of the campaign.