make wesnoth.simulate_combat a little more save against segfaults
it now keeps a unit_ptr to the unit to make sure the unit is not deleted
during whatever lua code the battle_context constructor calls. This can
be done because since the unit ctor refactor all unit objects are owned
by a unit_ptr (created by unit::create).

It might also be possible to refactor out the "unit * c_ptr" member of
lua_unit.