Fix #3650: rare crash in attack prediction
The calculation for the probability to be slowed was incorrect if both
units had slow and neither of them was slowed. Calculating the sum of all
probabilities in the "not hit" row/column is a valid approach: however,
if the unit itself can slow and the other unit isn't already slowed, then
some of the "not hit" cases end up into a different plane.

Probabilities from that plane are now counted as well.