Tutorial S2: Don't allow the orc to recruit wolves on turn 1
If the wolves are recruited on the first turn then Galdrad's turn 3 hint about
getting to the island before them is too difficult for a tutorial, and his hint
about holding the village on turn 4 hint may be a turn too late.

Some have to be recruited on turn 2 to fit the turn 3 hint about them, so rig
the AI's recruitment to do that.