small connect_engine cleanup
The only attributes that were changed by side_engine::new_config() and
not in the attributes_to_trim list were "user_description" and "name" so
it is cleaner to just explicitly change them than to blacklist
everything else and do a merge.

Also with this it no longer merges subtags, which in particular fixes
Issue #3338 where the [ai] subtags were messed up.

This is a backport of e036cadf0b