Help: cut portraits of monsters, bats and zombies
This will less often wrap the text of the descriptions.

For many of the zombies is more possible, but it has been taken care to have the
linebreak for 'required xp' and in the variation list the same on all their pages.

The reason for the 319px height on some of them is:
- It's still taking enough space to have 'required xp' on a new line.
- It does for both the WC and Soulless not interfere with the description text.
- The swimmer can't be smaller than 319px without cutting any non-tranaparent
  pixel. To have more structure the goblin/dwarf/dog use the same.

[ci skip]