AoI S1: Hint about recruiting a healer, if the player doesn't have any
This is just a hint on the general principle that it's good to have a healer.

When playing on easy, there's not much experience available in this first
scenario. Any hints about leveling units probably belong in S2 or later,
which also has the advantage of being able to talk about the enemies in S2.