clientside fix for #3717 : oos for allies caused by planned recruits
a servesided fix was already pushed, both fixes fix the issue independly: the issue will only appae when none of there fixes are present in the curent game.

The probem was that the wb network handler preiously might create units and the unit ctor would wrongly assumes it was in a synced context which messed up the unit id counter. I asume this happend because turn_info::process_network_data was called during the play_slice calls during an animation loop to process chat messages but actually received wb data.