AI move_to_targets CA: do not use dynamic limit for a* search
Using a dynamic limit here is not the right way to do this, among other things for the reasons given in the old comment. The new comment is copied from a bit earlier in the code, where more path finding is done. Path finding is indeed the most expensive part of the evaluation, but which of the two path finding loops takes more time depends on whether there are more goals or more units on the map.

In any case, if we want to improve the efficiency of this part of the code, it should be done by some sort of "smart pre-selection" rather than by limiting the a* search like this.  Also, this second path finding loop can already be skipped by using simple_targeting=yes.

Note that this commit does not change current behavior, it only removes the TODO and commented-out code.