AI: remove references to extended event system
There is no plan to implement anything like this. In fact, there is no information anywhere what this was even meant to be. My best guess is that it was some sort of log system of what actions the AI had done, or potentially was planning to do, so that actions could be considered in combination rather than in isolation. However, except for the recent_attacks table appearing right after these comments, nothing like this exists in the Wesnoth AI.

I am removing these comments in one commit separate from the TODO clean-up, so that this can be found more easily should it ever become relevant.