AI: remove invalidate_on_minor_gamestate_change
This has never been implemented and there is not even a clear use case for it. It was envisioned to be used to invalidate AI aspects when "minor" changes occurred that do not actually move an AI unit or change one of its main properties (such as changes to unit variables).