MP load game: do not overwrite saved AI
When an MP game is saved and then reloaded from the MP Create Game screen, one goes through the Game Lobby screen which gives the option of changing controllers and AIs for the different sides. Previously, this added an AI stage in front of the already existing AI for all computer player sides. This effectively replaced the existing AI.

This commit adds an option "Keep saved AI" to the Computer Player selection menu (when loading games only) and makes it the default. It does nothing to the AI config if this is chosen, otherwise it uses the new AI and also removes the old AI stages from the config.

This fixes #3791.