use pkg-config instead freetype-config
freetype-config has been removed from freetype earlier this year

compiling iwth gcc 4.3 works, and fails with >=4.5 and clang

At least on Archlinux one must provide the option --with-freetype-prefix=/usr/lib

As this is an old version of wesnoth, options such as --with-preferences-dir,
--program-suffix, --with-datadir-name and --docdir are useful to avoid collisions
with other wesnoth versions. For more, see ./configure --help.