wesnoth.find_path: deprecate old syntax for custom cost functions
The new syntax (using calculate= in the path options table) allows simultaneously setting the other path options, while there is no way to do so with the old syntax (passing the cost function as the last argument instead of the path options table).

We also remove the line allowing for an additional boolean argument to set ignore_teleports=. It does not make sense to single out this option from the others. Furthermore, this line makes it default to the opposite of the (more intuitive) default of the new syntax. For the record, this argument has not been in any released version yet and is not yet documented on the wiki.