campaignd: Explicitly check for before= and after= in add-ons list requests
A regression introduced in e2a64541002f0b4ee885f699c4f0901e441f938a
causes the timestamp-based list filtering to be forcibly enabled in ALL
cases because the author missed that the old code would bail out due to
lexical_cast() throwing an exception if the provided input was empty,
such as when before/after aren't provided in the list request. As a
result, the add-ons list would come up empty for clients because nothing
was matching the resulting time criteria.

The old pattern was a bit oblique to say the least, so I decided it'd be
best to replace it with an explicit check for the attributes' existence
and contents.