Fix item names in the scenario editor not being localized
data/core/editor/items.cfg defines the names for the items. The file is
included by data/core/editor/_main.cfg, which has a #textdomain binding
to wesnoth-editor at the top. The strings end up bound to that
textdomain by the game, but wmlxgettext doesn’t have any way to see this
so it binds them to the default textdomain instead, which is wesnoth,
not wesnoth-editor. The game ends up looking for the strings in
wesnoth-editor instead.

The easiest solution right now is to explicitly bind items.cfg to the
wesnoth textdomain so as to not require moving catalogue entries in the
po files around to wesnoth-editor. It really should've been
wesnoth-editor in the first place but it's too late to fix that now
without giving translators a migraine. It's a pity either way given how
useless the scenario editor is in practice.

[ci skip]