Fix wmlparser3 considers "_" part of attribute value
The wml parser, when parsing the outside of a string that is
translatable, only assumed “spaces” characters could precede it. If
there are “tabs” characters the code parses the “_” character as an
attribute value.

The solution is to add a stripping code for “tabs” characters.

Code style updated based on recommendations from ProditorMagnus.