Strip usernames used in the MP UI and command line of trailing whitespace
This also affects line breaks. There's no point in displaying confusing
errors to the player in the off-chance that they manage to introduce
leading or trailing whitespace in the username field. The server doesn't
accept whitespace but that doesn't mean we can't try to guess what the
user really meant to do, and extra whitespace around a name is a pretty
trivial mistake to make.

The password field, on the other hand, ought to accept it since it's
possible to even have line break characters in it (as tested with
Firefox on the live phpBB registration form for forums.w.o).