Minor documentation fix for config-dir and config-path
No string change, since this is the text for the output of
wesnoth --help, which isn't translated.

The manpages on the other hand only mention this option as being
deprecated, not specifing the details, so there is no change neccessary.

These options are an alias to userdata-dir/path, no more, no less.
Not affecting userconfig-dir/path (maybe thus the names are deprecated).

Issue #2000 remains