Fix parser opens tmp file 2 times then crashes
When parsing binary data in the wmlparser3 script, a temporary file is
created and opened. The problem is that the temporary file is never
closed and in subsequent functions it is reopened.

The solution is first, create a list of temporary file paths to delete at
program’s exit. In the atexit’s registered cleaning function, all the
files in this list are deleted.

Then in the function that creates the temporary file, we simply close
the file after the binary data has been written in it.

Closes #3927