Fix parser handles textdomain directives incorrectly
The problem is that when parsing #textdomain directives, the parser
expects the directive to be at the start of the line. So if the directive
is preceded by spaces or tabs then it is treated as an attribute.

The solution is to strip all spaces and tabs on the left side of the
parsed line string when checking for #textdomain directive.

Closes #3951