fixup! UI: Highlight enemy units in "Show Enemy Moves" mode (#3961)
In the test scenario, highlighting (26,12), pressing Ctrl+V,
highlighting (25,12), and pressing Ctrl+V again refreshed highlighting
of the Ghost but not of the Elvish Sharpshooter.  It was similar
with the enemy leaders in the first scenarios of AToTB and AOI.
I haven't figured out exactly why, but invalidating the units' hexes
like toggleellipse does seems to fix it.