Fix #2852: Reset statistics when rewinding a replay to the beginning.
Currently, while watching a replay of Scenario N, the statistics dialog
has the following options:

- All scenarios
- S1
- S2
- ...
- SN
- SN

The last option shows the statistics of Scenario N from the beginning
through the current point in the replay.  The other options show
end-of-scenario stats of each scenario and the cumulative stats of all
scenarios through the end of SN.

This patch causes the stats shown by the the second SN option to reset
to zero when the "reset replay" ("stop") button is selected.