DiD S09: Unblock the book if one of DV's units is standing on it
Refactor to using location_id for the location of the book, because it appears
multiple times in the .cfg file.

An edge case to this edge case: if DV's turn starts with DV next to the book
and one of DV's units on the book, DV doesn't move until the next turn.  This
was also true if the unit moved to attack something, and works in the player's
favor, as it gives Malin an extra turn to grab the book, so I'm treating it as
a feature.

If DV does grab the book, explicitly say in the objectives that it's now a race
to the escape tunnel. The sequence for that is:
* DV takes the book (no messages at all)
* silently update objectives (escape to win, same as Malin taking the book)
* player ends turn
* DV moves, Malin asks if he's going to be left behind
* start of player's next turn: objectives shown (race to win)