HttT S11: Correct two places where a typographic quote was used in Li'sar's id
In S11, if the player tries to attack Li'sar then there are supposed to be four
events adding extra guards. Two of them didn't trigger, because the [have_unit]
calls were trying to match her id with a typographic quotation mark.

One change in the opposite direction, where the displayed name could have a
non-typographic quotation mark; this might only happen when using choose_level
to reach a scenario with NEED_LISAR before her unit is on Konrad's recall list.

Remove a CLEAR_VARIABLE that was obsolete since 7778e6f4.