ANL: use loyal guardians and adjust income in return
Killing a guardian (the guards or a spider) punished the player,
because the AI then doesn't need to pay upkeep anymore for that unit.

That has been changed, and as the AI has thus less cost, the income needs
adjustments too:

The following has been taken into account:
- the guardian units had upkeep costs of 7g
- the AI has 3-4 villages, each worthy 1g of village_support
  * before turn 7, the AI can't recruit Lv 1 units. From turn 9 it recruits
    exclusively Lv 1 units.
  * This means in the early game the AI payed just 4g upkeep for the
    guardians and when it managed to recruit 3 Lv 1 units it was equal to 7g
- in the late game the AI may need more money in general, as Lv 2 units are
  expensive. The fact that the AI gets 50g in turn 17 means that the income
  is too low to buy the units normally.

The general idea is to have with this change same amount of money in the
early game and the same or more in the late game.

To address the above:
- 4g less income
- the Lv 3 spider is usually killed in each game, often before turn 7. It's
  seen as compensation for the village_gold
- The income of the AI is increased in relation to the turns. From turn 15 on:
  Each turn one of the AIs gets +1g income – in average +1g income each 4 turns.